# .github/workflows/swagger.yaml
openapi: 3.0.3
info:
  title: CargoHUB API
  description: API voor het beheren van klanten, orders, inventaris, locaties, zendingen, en transfers binnen CargoHUB.
  version: 1.0.0
servers:
  - url: https://localhost:5072/cargohub
    description: Productieomgeving
  - url: https://sandbox.api.cargohub.com
    description: Sandbox omgeving
paths:
...
  /item_groups/{item_group_id}/items:
    get:
      summary: Read all items for a specific item group
      parameters:
        - name: item_group_id
          in: path
          required: true
          description: The ID of the item group whose items are to be retrieved
          schema:
            type: integer
      responses:
        '200':
          description: Items found for the specified item group
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: No items found for the item group
          content:
            text/plain:
              schema:
                type: string
                example: "No items found for item_group with Id: {item_group_id}"
  /item_lines/{item_line_id}:
    get:
      summary: Read an item line by ID
      parameters:
        - name: item_line_id
          in: path
          required: true
          description: The ID of the item line to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Item line found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item_Line'
        '404':
          description: Item line not found
          content:
            text/plain:
              schema:
                type: string
                example: "No such item_line with Id: {item_line_id}"

    put:
      summary: Update an existing item line
      parameters:
        - name: item_line_id
          in: path
          required: true
          description: The ID of the item line to update
          schema:
            type: integer
      requestBody:
        description: Item line object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item_Line'
      responses:
        '200':
          description: Item line updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item_line updated successfully."
        '400':
          description: Failed to update item line
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update item_line."

    delete:
      summary: Delete an item line by ID
      parameters:
        - name: item_line_id
          in: path
          required: true
          description: The ID of the item line to delete
          schema:
            type: integer
      responses:
        '200':
          description: Item line deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item_line deleted successfully."
        '400':
          description: Failed to delete item line
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to delete item_line."
  /item_lines:
    get:
      summary: Retrieve all item lines
      responses:
        '200':
          description: Item lines found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item_Line'
        '404':
          description: No item lines found
          content:
            text/plain:
              schema:
                type: string
                example: "No item_lines found"

    post:
      summary: Create a new item line
      requestBody:
        description: Item line object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item_Line'
      responses:
        '200':
          description: Item line created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item_line created successfully."
        '400':
          description: Failed to create item line
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create item_line."
  /item_lines/{item_line_id}/items:
    get:
      summary: Read all items for a specific item line
      parameters:
        - name: item_line_id
          in: path
          required: true
          description: The ID of the item line whose items are to be retrieved
          schema:
            type: integer
      responses:
        '200':
          description: Items found for the specified item line
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: No items found for the item line
          content:
            text/plain:
              schema:
                type: string
                example: "No items found for item_line with Id: {item_line_id}"
...
  /item_types/{item_type_id}/items:
    get:
      summary: Read all items for a specific item type
      parameters:
        - name: item_type_id
          in: path
          required: true
          description: The ID of the item type whose items are to be retrieved
          schema:
            type: integer
      responses:
        '200':
          description: Items found for the specified item type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: No items found for the item type
          content:
            text/plain:
              schema:
                type: string
                example: "No items found for item_type with Id: {item_type_id}"
  /items/{item_id}:
    get:
      summary: Get an item by ID
      parameters:
        - name: item_id
          in: path
          required: true
          description: The ID of the item to retrieve.
          schema:
            type: string
      responses:
        200:
          description: Item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        404:
          description: Item not found
          content:
            text/plain:
              schema:
                type: string
                example: "No such item with Id: {item_id}"
    put:
      summary: Update an item by ID
      parameters:
        - name: item_id
          in: path
          required: true
          description: The ID of the item to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        200:
          description: Item updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item updated successfully."
        400:
          description: Failed to update item
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update item."
    delete:
      summary: Delete an item by ID
      parameters:
        - name: item_id
          in: path
          required: true
          description: The ID of the item to delete.
          schema:
            type: string
      responses:
        200:
          description: Item deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item deleted successfully."
        400:
          description: Failed to delete item
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to delete item."
  /items:
    get:
      summary: Get a list of all items
      responses:
        200:
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        404:
          description: No items found
          content:
            text/plain:
              schema:
                type: string
                example: "No items found"
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        200:
          description: Item created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Item created successfully."
        400:
          description: Failed to create item
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create item."
...
  /orders/{order_id}/items:
    get:
      summary: Get all items in an order
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to retrieve items for.
          schema:
            type: integer
      responses:
        200:
          description: Items found in order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        404:
          description: No items found in the order
          content:
            text/plain:
              schema:
                type: string
                example: "No items found in order with Id: {order_id}"
    put:
      summary: Update items in an order
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to update items in.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PropertyItem'
      responses:
        200:
          description: Items in the order updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Items in order updated successfully."
        400:
          description: Failed to update items in order
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update items in order."
...
  /shipments/{shipment_id}/items:
    get:
      summary: Get items of a shipment by ID
      parameters:
        - in: path
          name: shipment_id
          required: true
          schema:
            type: integer
          description: The ID of the shipment to retrieve items for.
      responses:
        '200':
          description: Items of the shipment found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PropertyItem'
        '404':
          description: No items found or shipment not found
    put:
      summary: Update items of a shipment by ID
      parameters:
        - in: path
          name: shipment_id
          required: true
          schema:
            type: integer
          description: The ID of the shipment to update items for.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PropertyItem'
      responses:
        '200':
          description: Shipment items updated successfully
        '400':
          description: Failed to update shipment items
...
  /suppliers/{supplier_id}/items:
    get:
      summary: Get items for a specific supplier by ID
      parameters:
        - in: path
          name: supplier_id
          required: true
          schema:
            type: integer
          description: The ID of the supplier to retrieve items for.
      responses:
        '200':
          description: Items for the supplier found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: No items found for the supplier or supplier not found
...
  /transfers/{transfer_id}/items:
    get:
      summary: Get items for a specific transfer by ID
      parameters:
        - in: path
          name: transfer_id
          required: true
          schema:
            type: integer
          description: The ID of the transfer to retrieve items for.
      responses:
        '200':
          description: Items for the transfer found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: No items found for the transfer or transfer not found
...
components:
  schemas:
...
    Item_Line:
      type: object
      properties:
        id:
          type: integer
          description: Unieke identificatie van de itemlijn
        name:
          type: string
          description: Naam van de itemlijn
        description:
          type: string
          description: Beschrijving van de itemlijn
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Item:
      type: object
      properties:
        uid:
          type: string
          description: Unieke ID van het item
        code:
          type: string
          description: Code van het item
        description:
          type: string
          description: Beschrijving van het item
        short_description:
          type: string
          description: Korte beschrijving van het item
        upc_code:
          type: string
          description: UPC code van het item
        model_number:
          type: string
          description: Modelnummer van het item
        commodity_code:
          type: string
          description: Commoditeit code van het item
        item_line:
          type: integer
          description: Itemlijn ID
        item_group:
          type: integer
          description: Itemgroep ID
        item_type:
          type: integer
          description: Itemtype ID
        unit_purchase_quantity:
          type: integer
          description: Aantal per aankoop eenheid
        unit_order_quantity:
          type: integer
          description: Aantal per bestel eenheid
        pack_order_quantity:
          type: integer
          description: Aantal per pack bestel eenheid
        supplier_id:
          type: integer
          description: Leverancier ID
        supplier_code:
          type: string
          description: Leverancier code
        supplier_part_number:
          type: string
          description: Leverancier onderdeelnummer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PropertyItem:
      type: object
      properties:
        id:
          type: string
          description: Item ID
        amount:
          type: integer
          description: Amount of items